requirements.txt
streamlit
google-cloud-aiplatform
Pillow